fix feature-type, allow truth to be draco
