+++ engine/runtime/examples/fixture/workflow/data-convert/plateau6/01-bldg/workflow.yml
@@ -102,6 +102,11 @@ graphs:
               value:
                 type: flowExpr
                 value: json.dumps(attributes["attributes"])
+            - attribute: __citygml_feature_type
+              method: rename
+              value:
+                type: string
+                value: feature_type
 
       - id: 9f2a7c4e-3b5d-4c8f-a1e0-6d9b2f5a8c31
         name: AttributeConversionTableForWardName
@@ -288,7 +293,6 @@ graphs:
         type: action
         action: Cesium 3D Tiles Writer
         with:
-          schemaKey: __citygml_feature_type
           targetTileSize: 1048576
           attachTexture: false
           skipUnexposedAttributes: true
@@ -350,7 +354,6 @@ graphs:
         type: action
         action: Cesium 3D Tiles Writer
         with:
-          schemaKey: __citygml_feature_type
           targetTileSize: 1048576
           attachTexture: true
           skipUnexposedAttributes: true
@@ -363,7 +366,6 @@ graphs:
         type: action
         action: Cesium 3D Tiles Writer
         with:
-          schemaKey: __citygml_feature_type
           targetTileSize: 1048576
           textureCodec: Untextured
           skipUnexposedAttributes: true
@@ -425,7 +427,6 @@ graphs:
         type: action
         action: Cesium 3D Tiles Writer
         with:
-          schemaKey: __citygml_feature_type
           targetTileSize: 1048576
           attachTexture: true
           skipUnexposedAttributes: true
@@ -438,7 +439,6 @@ graphs:
         type: action
         action: Cesium 3D Tiles Writer
         with:
-          schemaKey: __citygml_feature_type
           targetTileSize: 1048576
           textureCodec: Untextured
           skipUnexposedAttributes: true
