+++ engine/testing/data/testcases/data-convert/plateau6/01-bldg/ward/truth/cesium_lod2_truth.json
@@ -46,6 +46,7 @@
     "city_code": "13104",
     "city_max_lod": "2",
     "core_creationDate": "2024-03-15T00:00:00",
+    "feature_type": "bldg:Building",
     "gml_id": "bldg_77ca1a15-3b35-4386-8f86-152ed71c4c64",
     "lod": "2",
     "lodMask": "7",
