+++ engine/testing/plateau-tiles-test/src/bin/generate_truth.rs
@@ -3,7 +3,7 @@ use plateau_tiles_test::conv::cesium_statistics;
 use plateau_tiles_test::conv::mvt;
 use plateau_tiles_test::conv::mvt_png;
 use plateau_tiles_test::conv::raster3d;
-use plateau_tiles_test::file::{extract_zip_to_tmp, zip_dir};
+use plateau_tiles_test::file::{decompress_glbs, extract_zip_to_tmp, zip_dir};
 use plateau_tiles_test::profile_config::Convs;
 use serde::Deserialize;
 use std::fs;
@@ -104,6 +104,7 @@ fn run(profile_path: &Path) -> Result<(), String> {
             .expect("convs.cesium_attributes path must have a file name");
         let zip_path = truth_dir.join(stem).with_extension("zip");
         let tmp_dir = extract_zip_to_tmp(&zip_path)?;
+        decompress_glbs(&tmp_dir);
         let output_path = truth_dir.join(&entry.truth_path);
         let result = conv_cesium::write_cesium_json(&tmp_dir, &output_path, entry.casts.as_ref());
         fs::remove_dir_all(&tmp_dir).ok();
@@ -124,6 +125,7 @@ fn run(profile_path: &Path) -> Result<(), String> {
             .expect("convs.raster3d path must have a file name");
         let zip_path = truth_dir.join(stem).with_extension("zip");
         let tmp_dir = extract_zip_to_tmp(&zip_path)?;
+        decompress_glbs(&tmp_dir);
 
         let png_dir = truth_dir.join(&entry.truth_path);
         let _ = fs::remove_dir_all(&png_dir);
