+++ engine/testing/plateau-tiles-test/src/main.rs
@@ -2,7 +2,7 @@ use plateau_tiles_test::conv::cesium as conv_cesium;
 use plateau_tiles_test::conv::mvt;
 use plateau_tiles_test::conv::mvt_png;
 use plateau_tiles_test::conv::raster3d as conv_raster3d;
-use plateau_tiles_test::file::{extract_dir, zip_dir};
+use plateau_tiles_test::file::{decompress_glbs, extract_dir, zip_dir};
 use plateau_tiles_test::profile_config::Convs;
 use plateau_tiles_test::runner;
 use plateau_tiles_test::tester::cesium::{self, CesiumConfig};
@@ -25,7 +25,6 @@ use std::sync::Once;
 use tracing::info;
 use tracing_subscriber::prelude::*;
 use tracing_subscriber::EnvFilter;
-use walkdir::WalkDir;
 
 static INIT: Once = Once::new();
 
@@ -264,8 +263,6 @@ fn run_testcase(testcases_dir: &Path, results_dir: &Path, name: &str, stages: &s
         let flow_extracted_dir = output_dir.join("flow_extracted");
         extract_dir(&flow_source_dir, &flow_extracted_dir).unwrap();
 
-        // Decompress draco-compressed glb in flow output
-        // truth zips should be preprocessed to contain only decompressed glb files
         decompress_glbs(&flow_extracted_dir);
 
         let tests = &profile.tests;
@@ -468,32 +465,6 @@ fn run_testcase(testcases_dir: &Path, results_dir: &Path, name: &str, stages: &s
     }
 }
 
-fn decompress_glbs(flow_extracted_dir: &Path) {
-    let glb_files: Vec<_> = WalkDir::new(flow_extracted_dir)
-        .into_iter()
-        .filter_map(|e| e.ok())
-        .filter_map(|entry| {
-            let path = entry.path();
-            if path.is_file() && path.extension().is_some_and(|e| e == "glb") {
-                Some(path.to_path_buf())
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    let mut cmd = std::process::Command::new("glb-decompress");
-    for glb_file in &glb_files {
-        cmd.arg(glb_file.as_os_str());
-    }
-
-    let status = cmd
-        .status()
-        .expect("Failed to execute glb-decompress command");
-    if !status.success() {
-        panic!("glb-decompress failed");
-    }
-}
 
 fn main() {
     init_logging();
